minor: moved code related to issuing cards into 'helpers.py' module. -@iamserda
